refactor: DataRow の int ID 型を long に統一
SQLite の INTEGER 型は C# では long となるため、
ProductID / SubstrateID 系のプロパティ・メソッドシグネチャを
int から long に変更し、不要な (int) キャストを除去した。

- Models: ProductMaster.ProductID, SubstrateMaster.SubstrateID,
  SubstrateInfo.SubstrateID, ProductRegisterWork.ProductID,
  SubstrateRegisterWork.SubstrateID を long に変更
- ProductRepository: Insert/Delete/Exists 系メソッドの戻り値・引数を long に変更
- 各 Window/Dialog: ListItem<int> -> ListItem<long>、
  DataRow 読み取りの (int) キャストを除去
- Field<T>() の型を SQLite の実型と一致させ、変換の往復を排除

+++ ProductDataBase/MasterManagement/SubstrateMasterEditDialog.cs
@@ -24,7 +24,8 @@ namespace ProductDatabase.MasterManagement {
             if (!_isNewRecord && _sourceRow != null) {
                 _substrate.LoadFrom(_sourceRow);
                 PopulateFields();
-            } else {
+            }
+            else {
                 // 新規追加時のデフォルト値
                 VisibleCheckBox.Checked = true;
             }
@@ -132,7 +133,7 @@ namespace ProductDatabase.MasterManagement {
             }
 
             // 基板型式の重複チェック
-            int excludeId = _isNewRecord ? 0 : _substrate.SubstrateID;
+            long excludeId = _isNewRecord ? 0 : _substrate.SubstrateID;
             if (_repository.ExistsSubstrateModel(SubstrateModelTextBox.Text.Trim(), excludeId)) {
                 MessageBox.Show(
                     $"基板型式 [{SubstrateModelTextBox.Text.Trim()}] は既に登録されています。",
