+++ ProductDataBase/Models/SubstrateInfo.cs
@@ -1,6 +1,6 @@
 namespace ProductDatabase.Models {
     public class SubstrateInfo {
-        public int SubstrateID { get; set; }
+        public long SubstrateID { get; set; }
         public string SubstrateName { get; set; } = string.Empty;
         public string SubstrateModel { get; set; } = string.Empty;
     }
